Consistently remove federation link
Previously when a user successfully authenticated through the
legacy provider REST API, the Keycloak user would have the federation
link removed.

However, when the user account received a credential through any other
mechanism (forgot password flow, set in Keycloak UI by a user with
manage user permission, etc) the federation link would not be removed.

This behavior led to inconsistent production data based on the
mechanism that set the user credential. This fix moves the link removal
to the CredentialInputUpdater capability interface. Now any time the
credential is set (regardless of method) and the federation link is
present, it will be consistently removed.